Created scripts dir to keep a run function. Starting to separate params from utils.
