+++ pyproject.toml
@@ -42,7 +42,8 @@ dependencies = [
     'numpy',
     'matplotlib',
     'scikit-learn',
-    'scipy>=1.9.3'
+    'scipy>=1.9.3',
+    'click'
 ]
 
 [project.optional-dependencies]
